Enemy tanks of the 3rd and 4th levels are added.

+++ Attack/FirstStage.h
@@ -97,7 +97,7 @@ void drawMap(String map[H], RenderWindow &app, Animation &a, double time)
 			else
 				continue;
 
-			a.sprite.setPosition(j * 32, i * 32);
+			a.sprite.setPosition(float(j * 32), float(i * 32));
 			app.draw(a.sprite);
 		}
 }
@@ -113,28 +113,28 @@ void drawForestAndIcons(String map[H], RenderWindow &app, Animation &a, Animatio
 			if (map[i][j] == 'F')
 			{
 				a.sprite.setTextureRect(IntRect(32, 32, 32, 32));
-				a.sprite.setPosition(j * 32, i * 32);
+				a.sprite.setPosition(float(j * 32), float(i * 32));
 				app.draw(a.sprite);
 			}
 
 			if (map[i][j] == 'R')
 			{
 				icons[0].sprite.setTextureRect(icons[0].frames[(int)icons[0].frame]);
-				icons[0].sprite.setPosition(j * 32, i * 32);
+				icons[0].sprite.setPosition(float(j * 32), float(i * 32));
 				app.draw(icons[0].sprite);
 			}
 
 			if (map[i][j] == 'U')
 			{
 				icons[1].sprite.setTextureRect(icons[1].frames[(int)icons[1].frame]);
-				icons[1].sprite.setPosition(j * 32, i * 32);
+				icons[1].sprite.setPosition(float(j * 32), float(i * 32));
 				app.draw(icons[1].sprite);
 			}
 
 			if (map[i][j] == 'C')
 			{
 				icons[2].sprite.setTextureRect(icons[2].frames[(int)icons[2].frame]);
-				icons[2].sprite.setPosition(j * 32, i * 32);
+				icons[2].sprite.setPosition(float(j * 32), float(i * 32));
 				app.draw(icons[2].sprite);
 			}
 		}
